refactor: split ComponentRegistry into base (rari_core) + Flight extension (rari), remove 47 duplicate deps
